Fix jenkins tests.
